SargonObjectCodable working
