+++ src/profile/v100/header/device_info.rs
@@ -95,21 +95,6 @@ impl HasSampleValues for DeviceInfo {
 }
 
 impl DeviceInfo {
-    /// Creates a new `DeviceInfo` from json in the form of `String`.
-    /// This is a temporarily exported method that allows wallet clients to
-    /// integrate Profile in steps.
-    ///
-    /// Should be replaced later with `Wallet`
-    pub fn new_from_json_string(json: impl AsRef<str>) -> Result<Self> {
-        let json = json.as_ref();
-        serde_json::from_str::<Self>(json).map_err(|_| {
-            CommonError::FailedToDeserializeJSONToValue {
-                json_byte_count: json.len() as u64,
-                type_name: "DeviceInfo".to_owned(),
-            }
-        })
-    }
-
     /// Creates a new `DeviceInfo` from json in the form of Vec<u8>.
     /// This is a temporarily exported method that allows wallet clients to
     /// integrate Profile in steps.
@@ -123,16 +108,6 @@ impl DeviceInfo {
                 type_name: "DeviceInfo".to_owned(),
             }
         })
-    }
-
-    /// Converts this `DeviceInfo` to json in the form of `String`
-    /// This is a temporarily exported method that allows wallet clients to
-    /// integrate Profile in steps.
-    ///
-    /// Should be replaced later with `Wallet`
-    pub fn to_json_string(&self) -> String {
-        serde_json::to_string(self)
-            .expect("JSON serialization of DeviceInfo should never fail.")
     }
 
     /// Converts this `DeviceInfo` to json in the form of `Vec<u8>`
@@ -207,13 +182,6 @@ mod tests {
     #[test]
     fn date_is_now() {
         assert!(SUT::new_iphone().date.year() >= 2023);
-    }
-
-    #[test]
-    fn json_string_roundtrip() {
-        let sut = SUT::sample();
-        let json_str = sut.to_json_string();
-        assert_eq!(sut, SUT::new_from_json_string(json_str).unwrap());
     }
 
     #[test]
